% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_func.R
\name{assign_func}
\alias{assign_func}
\title{Assign main and detailed functional group based on any PR2 clade}
\usage{
assign_func(
  x,
  assign = "clade",
  ref = DBc,
  sep = ";",
  empty_string = c(NA, "NA", "", " ")
)
}
\arguments{
\item{x}{a vector of taxonomy path OR a data-frame with taxonomy path in a column named "taxonomy" or "Taxonomy" OR a data-frame with column named after PR2 v5.0.0 taxonomic ranks}

\item{assign}{function to apply, either \code{\link{assign_clade}} or \code{\link{assign_path}}}

\item{ref}{the reference database for performing the functional assignment, either \code{\link{DBc}}, \code{\link{DBu}} or a manual database containing at least a 'taxonomy' column}

\item{sep}{separator character used to separate taxonomic ranks in the taxonomic path}

\item{empty_string}{a vector of element accepted for missing taxonomic information at a single taxonomic rank}
}
\value{
a data-frame composed of:
\itemize{
  \item the input taxonomy path 
  \item all other columns if `x` is a data-frame 
  \item assigned_from: the matched taxonomy path in the condensed database 
  \item assigned_at_rank: the deepest taxonomic rank name of the matched taxonomy path 
  \item assigned_from_taxa: the deepest clade name of the matched taxonomy path 
  \item all the columns of EukFunc database containing the functional informations
}
}
\description{
Generalized approach to assign a functional group to a taxonomic path using either the condensed clade or unique clade databases.
When a taxonomic path could not be assigned to any functional group, assignment is the made on the database version containing only the main_functional class.
}
\examples{
taxo_path <-
  gsub("[\r\n ]","",
    c("Eukaryota;TSAR;Alveolata;Ciliophora;Colpodea;
      Colpodea_X;Colpodida;Colpoda;Colpoda_cucullus;",
      "Eukaryota;TSAR;Alveolata;Ciliophora;Colpodea;
      Colpodea_X;Platyophryida;Platyophrya;Platyophrya_sp.",
      "Eukaryota;Obazoa;Opisthokonta;Fungi;Ascomycota;
      Pezizomycotina;Sordariomycetes;Fusarium;Fusarium_sp.;"))
assign_func(taxo_path)

# Using a data-frame with separated taxonomic rank columns as input:
path_func_exa <- assign_func(head(taxo_exa[10:19,]))
dplyr::count(path_func_exa, assigned_at_rank)
dplyr::count(path_func_exa, main_functional_class)

# Using a data-frame with a single taxonomy column:
path_func_exa <- assign_func(head(taxo_exa2))
dplyr::count(path_func_exa, assigned_at_rank)
dplyr::count(path_func_exa, main_functional_class)
}
\seealso{
\code{\link{assign_clade}}, \code{\link{assign_path}}, \code{\link{DBu}}, \code{\link{DBc}}, \code{\link{DBu_main}}, \code{\link{DBc_main}}
}
