% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_genbank.R
\name{assign_genbank}
\alias{assign_genbank}
\title{Assign functional group based on genbank accession}
\usage{
assign_genbank(x, ...)
}
\arguments{
\item{x}{a vector or list of genbank accessions}

\item{...}{further arguments passed to \code{\link{assign_path}} and \code{\link{consensing}}}
}
\value{
a data-frame with the input list of accessions and the outputs of \code{\link{consensing}} and \code{\link{assign_path}}
}
\description{
Read one or multiple genbank accessions, extract the corresponding taxonomy from the SoilEukFunc full database, compute consensus taxonomy if multiple accessions are provided and assign the corresponding functional group using \code{\link{assign_path}}
}
\examples{
acc<-c("AB918716.1,EU039893.1,JF747215.1",
       "EU039893.1, JF747215.1",
       "AB918716.1,EU039893.1,JF747215.1",
       "ABC")
assign_genbank(acc)
}
\seealso{
\code{\link{consensing}}, \code{\link{DBf}}, \code{\link{assign_path}}, \code{\link{assign_sp}}
}
