% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_majority.R
\name{assign_majority}
\alias{assign_majority}
\title{Improve functional assignment based on a partial PR2 taxonomic path and majority rule}
\usage{
assign_majority(
  x,
  threshold = 95,
  ref = DBc,
  sep = ";",
  empty_string = c(NA, "NA", "", " ")
)
}
\arguments{
\item{x}{a data-frame as produced by \code{\link{assign_path}}}

\item{threshold}{a species percentage threshold above which functional information of the majority of species in a clade are assigned to a taxonomic path}

\item{ref}{the condense database for performing the functional assignment, should be the same as used for \code{\link{assign_path}}, default to \code{\link{DBc}}}

\item{sep}{separator character used to separate taxonomic ranks in the taxonomic path}

\item{empty_string}{a vector of element accepted for missing taxonomic information at a single taxonomic rank}
}
\value{
a data-frame composed of:
\itemize{
  \item taxonomy: the input taxonomy path 
  \item all other columns of `x` not generated by \code{\link{assign_path}} 
  \item assigned_from: the matched taxonomy path in the condensed database 
  \item assigned_at_rank: the deepest taxonomic rank name of the matched taxonomy path 
  \item assigned_from_taxa: the deepest clade name of the matched taxonomy path 
  \item all the columns of the selected SoilEukFunc database containing the functional informations 
  \item confidence: the percentage of species with identical functional informations in the clade described by assigned_from_taxa 
}
}
\description{
Read a data.frame generated with \code{\link{assign_path}} and assign unassigned taxonomic path to function if a majority of species in the taxonomic path share the same functional information.
WARNING: this function may lead to wrong functional annotation(s), particularly in clades with low number of species in the database. Use with caution!
}
\details{
Using ref parameter, other database can be used to assign a function
}
\examples{
fe <- assign_path(taxo_exa2, ref=DBc_main)
fx <- assign_majority(fe, ref=DBc_main, threshold=90)
dplyr::count(fe, main_functional_class, secondary_functional_class)
dplyr::count(fx, main_functional_class, secondary_functional_class)
}
\seealso{
\code{\link{assign_path}}, \code{\link{DBc}}, \code{\link{DBc_main}}
}
