% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_path.R
\name{assign_path}
\alias{assign_path}
\title{Assign functional group based on a partial PR2 taxonomic path}
\usage{
assign_path(x, ref = DBc, sep = ";", empty_string = c(NA, "NA", "", " "))
}
\arguments{
\item{x}{a vector of taxonomy path OR a data-frame with taxonomy path in a column named "taxonomy" or "Taxonomy" OR a data-frame with column named after PR2 v4.12.0 taxonomic ranks}

\item{ref}{the reference database for performing the functional assignment, default to \code{\link{DBc}}}

\item{sep}{separator character used to separate taxonomic ranks in the taxonomic path}

\item{empty_string}{a vector of element accepted for missing taxonomic information at a single taxonomic rank}
}
\value{
a data-frame composed of:
\itemize{
  \item{}{taxonomy: the input taxonomy path}
  \item{}{all other columns if `x` is a data-frame}
  \item{}{assigned_from: the matched taxonomy path in the condensed database}
  \item{}{assigned_at_rank: the deepest taxonomic rank name of the matched taxonomy path}
  \item{}{assigned_from_taxa: the deepest clade name of the matched taxonomy path}
  \item{}{all the columns of SoilEukFunc database containing the functional informations}
}
}
\description{
Read a taxonomy path and assign the corresponding functional group in the condensed clade database by partial matches of the provided taxonomy onto the full database clade path
}
\details{
Using ref parameter, other database can be used to assign a function
}
\examples{
taxo_path <-
  c("Eukaryota;Alveolata;Ciliophora;Colpodea;Colpodea_X;Colpodida;Colpoda;Colpoda_cucullus;",
    "Eukaryota;Opisthokonta;Fungi;Ascomycota;Pezizomycotina;Sordariomycetes;Fusarium;Fusarium_sp.;")
assign_path(taxo_path)

# Using a data-frame with separated taxonomic rank columns as input:
path_func_exa <- assign_path(taxo_exa)
dplyr::count(path_func_exa, assigned_at_rank)
dplyr::count(path_func_exa, main_functional_class)

\dontrun{
# Using a data-frame with a taxonomy column as input:
path_func_exa2 <- assign_path(taxo_exa2)
dplyr::count(path_func_exa2, assigned_at_rank)
dplyr::count(path_func_exa2, main_functional_class)

# Improve the assignment using the condensed database only for the main functional class:
path_func_exa2 <- assign_path(taxo_exa2, ref=DBc_main)
dplyr::count(path_func_exa2, assigned_at_rank)
dplyr::count(path_func_exa2, main_functional_class)

# Compare with published results
dplyr::count(path_func_exa2, Functions, main_functional_class)
}
}
\seealso{
\code{\link{consensing}}, \code{\link{DBc}}, \code{\link{DBc_main}}
}
