% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_sp.R
\name{assign_sp}
\alias{assign_sp}
\title{Assign functional group at species level}
\usage{
assign_sp(x)
}
\arguments{
\item{x}{a vector of species names OR a data-frame with species names in a column named "species" or "Species"}
}
\value{
a data-frame composed of the input vector or columns, a new column `species_name` use to perform the match and all the columns of the species unique EukFunc database without taxonomy.
        Species names with no match return the EukFunc databases columns filled with NA.
}
\description{
Read a species list and assign the corresponding functional group in the unique species database by exact matches of species names
}
\examples{
data(taxo_exa)
sp_func_exa<-assign_sp(taxo_exa)
}
\seealso{
\code{\link{assign_path}}, \code{\link{assign_genbank}}
}
