% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBc_minimal.R
\docType{data}
\name{DBc_minimal}
\alias{DBc_minimal}
\title{Minimal condensed version of the EukFunc database.}
\format{
A data frame with 5538 rows and 5 variables:
\describe{
  \item{taxonomy}{Taxonomic path of clades for which a unique functional information is available}
  \item{main_functional_class}{Main nutrient uptake mode of a given taxon}
  \item{level}{Taxonomic rank name at which the functional assignment is condensed/unique}
  \item{taxa}{Clade name at which the functional assignment is condensed/unique}
  \item{count}{Amount of species grouped under the same functional assignment}
}
}
\usage{
data(DBc_minimal)
}
\description{
A database containing the condensed taxonomy of all soil-associated protists,
fungi and nematodes clades from PR2 v5.0.0 and
their unique main functional class from the EukFunc database.
}
\seealso{
\code{\link{DBu}}, \code{\link{DBc}}, \code{\link{DBc_main}}
}
\keyword{datasets}
