% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilEukFuncDev.R
\docType{package}
\name{EukFunc}
\alias{EukFunc}
\title{Eukaryotic Soil Functional assignment tools and database}
\description{
Assign functional information to terrestrial protist, fungi and nematodes.
}
\details{
Provide the database of functional assignment at sequence, species and aggregated clade levels
}
\author{
Guillaume Lentendu \email{guillaume.lentendu@unine.ch}
}
