% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbio_det.R
\name{get_symbio_det}
\alias{get_symbio_det}
\title{Extract and format intermediate functional class information from symbiotroph detailed functional information}
\usage{
get_symbio_det(x)
}
\arguments{
\item{x}{a string from "detailed_functional_class" or "detailed_secondary_functional_class"}
}
\value{
a string with intermediate symbiotroph class among parasite, mycorrhiza, host phototroph and other
}
\description{
Extract and format intermediate functional class information from symbiotroph detailed functional information
}
\keyword{internal}
