% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensing.R
\name{consensing}
\alias{consensing}
\title{Compute consensus taxonomy}
\usage{
consensing(x, cons = 60, sep_string = ",", no_match = "no_match")
}
\arguments{
\item{x}{a vector or list of comma separated genbank accessions}

\item{cons}{consensus threshold in percent, default to 60}

\item{sep_string}{separator for multiple accessions in one element, default to comma}

\item{no_match}{value to search for for an absence of math, default to "no_match"}
}
\value{
a data-frame with the provided accessions, the assign taxonomy path, the assign clade and the assign taxonomic rank
}
\description{
A consensus taxonomy is created from multiple Genbank accessions matching PR2 accessions present in the EukFunc full database
}
\examples{
acc<-c("AB918716.1,EU039893.1,JF747215.1",
       "EU039893.1, JF747215.1",
       "AB918716.1,EU039893.1,JF747215.1",
       "ABC")
consensing(acc)

\dontrun{
# Compare to already assigned taxonomy:
cons_exa <- consensing(taxo_exa$reference)
cbind(input=tidyr::unite(taxo_exa, taxonomy, 2:9, sep=";")$taxonomy,
      dplyr::select(cons_exa, assigned_taxonomy, matched_accession, assigned_consensus))
}

}
\seealso{
\code{\link{DBf}}
}
